use uint256

+++ contracts/RemixChallenges.sol
@@ -22,8 +22,8 @@ contract RemixChallenges is Initializable, AccessControlUpgradeable, UUPSUpgrade
     uint256 public challengeIndex;
     mapping  (uint256 => Challenge) public challenges;
 
-    mapping (bytes => uint) public nullifiers;
-    mapping (bytes => uint) public publishers;
+    mapping (bytes => uint256) public nullifiers;
+    mapping (bytes => uint256) public publishers;
 
     address public rewardContract;
     
@@ -64,7 +64,7 @@ contract RemixChallenges is Initializable, AccessControlUpgradeable, UUPSUpgrade
       * @param proof The struct Proof containing all necessary data for a proof. This includes values a, b, c and d.
       * @param input The array of 3 uints containing additional input data for a challenge. This includes a challengeHash, a random number (r) and a nullifier (s).
       */
-    function publishChallenge (uint256 index, ZKVerifier.Proof memory proof, uint[3] memory input) public {
+    function publishChallenge (uint256 index, ZKVerifier.Proof memory proof, uint256[3] memory input) public {
         require(rewardContract != address(0), "reward contract not set");
         Challenge storage challenge = challenges[index];
         require(challenge.set == 1, "challenge not set");
